Do not allow multiple gpio refs in configure_*_gpios().
bluetooth_platform_data.bluetooth_power() is called for every write to the
bluetooth rfkill sys file, which in turn results in an arbitrary number of
calls to configure_*_gpios().  The gpiomux objects are reference counted,
so we cannot allow arbitrary gets and puts.

Fix this by only allowing one gpiomux reference in configure_*_gpios().

Change-Id: I0184821e3c5c907f1bcb246cd7766e52ebefdacf